spotify-explorer/requirements.txt
spotipy>=2.23.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tabulate>=0.9.0
